ghost/pyproject.toml
[project]
name = "ghost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click",
    "pyyaml",
    "gymnasium",
    "numpy",
    "pytest",
]

[project.scripts]
ghost = "cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/cli",
    "src/agent",
    "src/envs",
    "src/engine",
]
